Refactor: Split useLibrary.ts into sub-hooks and resolve ESLint line limit violations (#513)
