+++ eslint.config.mjs
@@ -133,7 +133,6 @@ export default ts.config(
       "src/components/organisms/MintingView.tsx",
       "src/components/organisms/ParameterEditor.tsx",
       "src/components/organisms/Workbench.tsx",
-      "src/hooks/useLibrary.ts",
       "src/lib/backup-validator.ts",
       "src/lib/db.ts",
       "src/lib/export-utils.ts",
@@ -154,7 +153,6 @@ export default ts.config(
       "src/components/templates/SidePanelLayout.tsx",
       "src/contents/_domain/extractors/WebDataExtractor.ts",
       "src/hooks/useDragAndDrop.ts",
-      "src/hooks/useLibrary.ts",
       "src/hooks/usePromptInjector.ts",
       "src/hooks/useSpotlight.ts",
       "src/lib/backup-validator.ts",
@@ -222,7 +220,6 @@ export default ts.config(
       "src/hooks/useEasyModeView.ts",
       "src/hooks/useEvolution.ts",
       "src/hooks/useExpertModeView.ts",
-      "src/hooks/useLibrary.ts",
       "src/hooks/useLocalBackup.ts",
       "src/hooks/usePromptInjector.ts",
       "src/hooks/useSpotlight.ts",
